tf66 forbid deleting creator

+++ src/main/java/com/theflow/service/UserServiceImpl.java
@@ -14,9 +14,11 @@ import java.util.List;
 import java.util.Set;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import validation.CompanyAliasExistsException;
+import validation.CompanyCreatorDeletingException;
 import validation.CompanyExistsException;
 import validation.EmailExistsException;
 import validation.UsernameDuplicationException;
@@ -37,6 +39,9 @@ public class UserServiceImpl implements UserService {
     
     @Autowired
     private UserCompanyDao userCompanyDao;
+    
+    @Autowired
+    private PasswordEncoder passwordEncoder;
 
     //Saves user from registration page. Assignes admin role
     @Override
@@ -116,7 +121,7 @@ public class UserServiceImpl implements UserService {
         user.setFirstName(userDto.getFirstName());
         user.setLastName(userDto.getLastName());
         user.setEmail(userDto.getEmail());
-        user.setPassword(userDto.getPassword());
+        user.setPassword(passwordEncoder.encode(userDto.getPassword()));
         Company company = companyDao.getCompanyByName(userDto.getCompanyName());
         user.setEnabled(true);
         
@@ -139,7 +144,11 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public void removeUser(int id) {
+    public void removeUser(int id) throws CompanyCreatorDeletingException {
+        Company company = companyDao.getCompanyById(getPrincipal().getCompanyId());
+        if (company.getCreator().getUserId() == id) {
+            throw new CompanyCreatorDeletingException("Illegal user deletion");
+        }
         userDao.removeUser(id, getPrincipal().getCompanyId());
     }
